- moves refiners to a dedicated namespace
