+++ src/Kiota.Builder/Refiners/JavaRefiner.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Kiota.Builder.Extensions;
 
-namespace Kiota.Builder {
+namespace Kiota.Builder.Refiners {
     public class JavaRefiner : CommonLanguageRefiner, ILanguageRefiner
     {
         public override void Refine(CodeNamespace generatedCode)
